+++ billreminder/model/schemas.py
@@ -80,7 +80,7 @@ class BillSchema(ModelSchema):
     id = fields.Int(dump_only=True)
     name = fields.String(required=True)
     description = fields.String(required=True)
-    amount = fields.Integer(required=True, validate=lambda n: n >= 0)
+    amount = fields.Float(required=True, validate=lambda n: n >= 0)
     last_payment = fields.DateTime(allow_none=True, format=DATE_FORMAT)
     participants = ma.List(ma.Nested(UserSchema))
     payments = ma.List(ma.Nested(PaymentSchema))
